move deduplicate functions into dedicated module
