+++ tests/consistify_test.py
@@ -21,6 +21,21 @@ class TestConsistify(unittest.TestCase):
         nptesting.assert_array_equal(wgs_consist, test_wgs_consist)
         nptesting.assert_array_equal(cattle_consist, test_cattle_consist)
         nptesting.assert_array_equal(movement_consist, test_movement_consist)
+        # test that dupliactes entires are not removed
+        test_wgs = pd.DataFrame({"Submission": ["A", "A"]})
+        test_cattle = pd.DataFrame({"CVLRef": ["A"]})
+        test_movement = pd.DataFrame({"SampleName": ["A"]})
+        test_wgs_consist = pd.DataFrame({"Submission": ["A", "A"]})
+        test_cattle_consist = pd.DataFrame({"CVLRef": ["A"]})
+        test_movement_consist = pd.DataFrame({"SampleName": ["A"]})
+        # run consistify
+        (wgs_consist, cattle_consist, movement_consist, *_) = \
+            consistify.consistify(test_wgs, test_cattle, test_movement)
+        # assert output
+        nptesting.assert_array_equal(wgs_consist, test_wgs_consist)
+        nptesting.assert_array_equal(cattle_consist, test_cattle_consist)
+        nptesting.assert_array_equal(movement_consist, test_movement_consist)
+
 
     def test_clade_correction(self):
         # test input
