+++ tests/unit_tests.py
@@ -5,6 +5,7 @@ from phylogeny_test import TestPhylogeny
 from filter_samples_test import TestFilterSamples
 from update_summary_test import TestUpdateSummary
 from consistify_test import TestConsistify
+from de_duplicate_test import TestDeDuplicate
 from utils_test import TestUtils
 
 
@@ -18,11 +19,11 @@ if __name__ == "__main__":
                       TestPhylogeny('test_match_s3_uri'),
                       TestPhylogeny('test_process_sample_name'),
                       TestPhylogeny('test_post_process_snps_df')]
-    filter_samples_test = [TestFilterSamples('test_remove_duplicates'),
-                           TestFilterSamples('test_get_indexes_to_remove'),
-                           TestFilterSamples('test_filter_df'),
+    filter_samples_test = [TestFilterSamples('test_filter_df'),
                            TestFilterSamples('test_filter_columns_numeric'),
                            TestFilterSamples('test_filter_columns_categorical')]
+    de_duplicate_test = [TestDeDuplicate('test_remove_duplicates'),
+                         TestDeDuplicate('test_get_indexes_to_remove')]
     update_summary_test = [TestUpdateSummary('test_append_df_summary'),
                            TestUpdateSummary('test_get_finalout_s3_keys'),
                            TestUpdateSummary('test_extract_s3_key')]
@@ -32,7 +33,7 @@ if __name__ == "__main__":
     runner = unittest.TextTestRunner()
     parser = argparse.ArgumentParser(description='Test code')
     module_arg = parser.add_argument('--module', '-m', nargs=1, 
-                                     help="module to test: phylogeny, update_summary, filter_samples, consistify or uitls'",
+                                     help="module to test: phylogeny, update_summary, filter_samples, de_duplicate, consistify or uitls'",
                                      default=None)
     args = parser.parse_args()
     if args.module:
@@ -40,6 +41,8 @@ if __name__ == "__main__":
             runner.run(test_suit(phylogeny_test)) 
         elif args.module[0] == 'filter_samples':
             runner.run(test_suit(filter_samples_test)) 
+        elif args.module[0] == 'de_duplicate':
+            runner.run(test_suit(de_duplicate_test)) 
         elif args.module[0] == 'update_summary':
             runner.run(test_suit(update_summary_test)) 
         elif args.module[0] == 'consistify':
